refactor(app): rename common/ui to common/components
Aligns naming with the rest of the app (features use _components/).
Updates the import in site-header and the two CLAUDE.md mentions.
